+++ app/(route)/(auth)/login/_components/LogInForm/index.tsx
@@ -4,8 +4,9 @@ import { useState } from 'react';
 import type { FieldError, RegisterOptions } from 'react-hook-form';
 import { useForm } from 'react-hook-form';
 import { useRouter } from 'next/navigation';
-import type { LoginFormValues } from '@/_apis/type';
 import useUserStore from '@/store/useUserStore';
+
+import type { LoginFormValues } from '@/_types/authentication';
 
 import { useModal } from '@/_hooks/useModal';
